Added searchbar logic to work with redux. Renamed files to keep same pattern
