+++ src/features/locations/table/locations-table.component.tsx
