+++ src/features/modal/index.ts
@@ -1,3 +1,3 @@
-import useDisclosure from './useDisclosure';
+import useDisclosure from 'features/modal/use-disclosure';
 
 export { useDisclosure };
