+++ src/features/navigation/index.ts
@@ -1,3 +1,3 @@
-import Navbar from 'features/navigation/Navbar';
+import Navbar from 'features/navigation/navigation.component';
 
 export { Navbar };
