+++ src/features/navigation/mobile/index.ts
@@ -1,3 +1,3 @@
-import NavMobile from './NavMobile';
+import NavMobile from 'features/navigation/mobile/navigation-mobile.component';
 
 export { NavMobile };
